Change size of sprites and modifying .png files

+++ headers/map.h
@@ -2,7 +2,7 @@
 char *map();
 SDL_Surface *load_Pathed_Media_Map(char *path, float scale);
 SDL_Surface *load_Background_Media_Map(char *background_image);
-SDL_Surface *loadMedia(char *path, int scale);
+SDL_Surface *loadMedia(char *path, float scale);
 void renderMap(SDL_Surface *map, int x, int y, int w, int h);
 void renderMapText(SDL_Surface *surface, char *Text, int font_size, float width_divide, float height_divide);
 int generateMaps(int day, char map_choices[][20], char countries[][20], int map_occurrence, char map_chosen[][20], SDL_Surface *maps[]);
